feat(Commu): postEditer Style 분리
